requirements.txt
Django>=1.5
django-extensions==1.3.3
six==1.5.2
wsgiref==0.1.2
djangorestframework>=2.3
django-jsonschema==0.2.0
